+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.8.0] - 2025-06-04
+### Added
+- [Marek Kurańda](https://github.com/mjkuranda): `KeywordService` that calculates `priorityScore`.
+
+### Changed
+- [Marek Kurańda](https://github.com/mjkuranda): The way how priority score is calculating.
+- [Marek Kurańda](https://github.com/mjkuranda): Index only on `priority` column in the database.
+
 ## [1.7.0] - 2025-06-03
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): `LoggerModule` to log each event within the system.